dj-sso/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Deprecated==1.2.14
Django==4.0.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jwcrypto==1.5.0
Markdown==3.4.3
oauthlib==3.1.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
